requirements.txt
openai 
sentence-transformers 
faiss-cpu 
hf_xet 
pdfplumber
python-docx
pytest
PyPDF2
pypdf
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.7.0,<3.0.0
python-multipart==0.0.6